+++ client/src/App.scss
@@ -7,6 +7,7 @@
 @import './stylesheets/Notfound';
 @import './stylesheets/SignUp';
 @import './stylesheets/BlogCreate';
+@import './stylesheets/MyPage';
 
 .main {
   margin: 0 auto;
